refactor language spefic service and add support to autoFormatter
